File class updates

+++ www/app/helper/files.php
@@ -2,22 +2,16 @@
 
 
 function deleteDirectory($dir) {
-    if (!file_exists($dir)) {
-        return true;
-    }
+    
+    if ( !file_exists($dir) ) return true;
 
-    if (!is_dir($dir)) {
-        return unlink($dir);
-    }
+    if ( !is_dir($dir) ) return unlink($dir);
 
     foreach (scandir($dir) as $item) {
-        if ($item == '.' || $item == '..') {
-            continue;
-        }
 
-        if (!deleteDirectory($dir . DIRECTORY_SEPARATOR . $item)) {
-            return false;
-        }
+        if ($item == '.' || $item == '..') continue;
+
+        if ( !deleteDirectory($dir . DIRECTORY_SEPARATOR . $item) ) return false;
 
     }
